code - shahaf draft
Objective:
a table in Snowflake that has information about the IaC State files providers, resources, and accounts (depends on info in Elastic). This is for us to understand which providers are in use so we can prioritize and gain insights.

The way to do it is a python script in the Product Cluster.